#664: Fix show-if/hide-if and tab/tab-group inline edition
